requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.21.0

# Computer Vision
opencv-python>=4.7.0
Pillow>=9.0.0
# The detectron2 package must be installed separately via git+https
# See README for installation instructions

# Web Application
streamlit>=1.22.0

# Image Processing & Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
requests>=2.28.0
tqdm>=4.64.0
pandas>=1.5.0
scipy>=1.9.0

# Configuration & Logging
pyyaml>=6.0
# logging is part of Python standard library
# argparse is part of Python standard library

# Optional: For advanced features
scikit-image>=0.19.0
imageio>=2.22.0
